requirements.txt
asgiref==3.4.1
Django==3.2.9
psycopg2-binary==2.9.1
pytz==2021.3
sqlparse==0.4.2
sgiref==3.4.1
autopep8==1.5.7
cffi==1.14.6
cryptography==3.4.7
Django==3.2.5
django-cors-headers==3.8.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
mysql-connector-python==8.0.25
protobuf==3.17.3
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
pytz==2021.1
six==1.16.0
toml==0.10.2